✨feat: Store 등록, 수정시 필요한 위도, 경도의 제한 범위 검증 로직 추가하기

+++ src/main/java/com/study/demo/backend/domain/store/converter/StoreConverter.java
@@ -11,8 +11,8 @@ public class StoreConverter {
     public static Store toEntity(StoreReqDTO.Create dto) {
         return Store.builder()
                 .name(dto.name())
-                .latitude(BigDecimal.valueOf(dto.latitude()))
-                .longitude(BigDecimal.valueOf(dto.longitude()))
+                .latitude(dto.latitude())
+                .longitude(dto.longitude())
                 .openingTime(dto.openingTime())
                 .closingTime(dto.closingTime())
                 .build();
